A few misc. fixes
-Added test suite AllTests which runs all test cases.  We can use this
to generate our EclEmma Coverage report.

-Had the board configuration arrays "upside down" in class TestBoard,
fixed this
-A few minor tweaks to the comments in class TestBoard

-The board configuration arrays in class TestGame were really messed up,
I had them so they had the rows/columns flipped (I think that's what you
were seeing earlier, Carlos) and when I thought I had the chips at the
bottom of the board it was actually the side.  Regardless, I fixed this,
and it should be fine now.