Implementing Bugfixes #1
Implemented fixes for all defects on Master Defect List except for
defect #2, defect #9 (which is a duplicate of defect #2) and defect #10.
For more details, see the Master Defect List under Deliverable 7 folder
on Google Drive.

In short, I think I've fixed all the defects we identified except for
the defect regarding the program crashing if a game chip is dropped into
an already-full column.

I think the game is much more usable now - all the prompts have been
improved, I modified the board-printing function so that the board
display is clearer, draws are now reported, game winner and score
display is much improved, and game restart functionality is now much
smoother.