Implementing Bugfixes #2
Fixed the bug where the program crashes if you attempt to place a game
piece in an already full column (defect #2 on the Master Defect List).
Now, instead of crashing, it informs the player that that is an invalid
choice and re-prompts them to enter another column.

With this, I believe all the defects on our Master Defect List are now
fixed, with the sole exception of defect #10.  I'm not sure if defect
#10 will wind up being a problem, so I've gone ahead and exported this
version of the program to an executable .jar file.  You can find it in
this repository in the folder "Executable jar files", it is named
Group6_Connect4_V1.jar